refactor: refactored app layout added, renamed and removed components
